+++ src/imgtools/__init__.py
@@ -3,7 +3,7 @@ __version__ = "2.0.0-rc.8"
 from .coretypes import BoxPadMethod, Coordinate3D, RegionBox, Size3D, Spacing3D
 from .datasets import example_data, example_data_paths
 from .dicom import find_dicoms, lookup_tag, similar_tags, tag_exists
-from .logging import logger
+from .loggers import logger
 
 __all__ = [
     "find_dicoms",
