+++ src/imgtools/dicom/dicom_find.py
@@ -4,7 +4,7 @@ from typing import List
 
 from pydicom.misc import is_dicom
 
-from imgtools.logging import logger
+from imgtools.loggers import logger
 
 
 def _is_valid_dicom(file: Path, check_header: bool) -> bool:
