+++ tests/conf_helpers.py
@@ -5,7 +5,7 @@ from typing import Union
 
 import pytest
 
-from imgtools.logging import logger  # type: ignore
+from imgtools.loggers import logger  # type: ignore
 
 ####################################################################################################
 # here we create some helpers for the dataset downloading
